requirements.txt
python-docx
fastapi[standard]
python-dotenv
llama-index
llama-index.core
llama_index.llms.bedrock
openai
llama_index.llms.azure_openai
llama_index.embeddings.azure_openai
openpyxl
docx2txt